+++ src/index.html
@@ -8,12 +8,12 @@
 </head>
 <body>
 <header>
-    <h1>World Regions</h1>
+    <a href="search.html"><h1>World Regions</h1></a>
     <img id="world-image" src="assets/world_map.png" alt="image of the world">
 </header>
 <main>
     <ul id="country-list"></ul>
 </main>
-<script type="module" src="app.js"></script>
+<script type="module" src="index.js"></script>
 </body>
 </html>
