+++ src/search.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <link rel="stylesheet" href="styles.css">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Find a country</title>
+</head>
+<body>
+<header>
+    <a href="index.html"><h1>World Regions</h1></a>
+    <img id="world-image" src="assets/screenshot2.png" alt="image of the world">
+</header>
+<main>
+    <form action="#">
+        <input type="text" id="search-input" placeholder="Hello World">
+        <span><input type="submit"></span>
+        <div id="country-info"></div>
+    </form>
+</main>
+<script type="module" src="search.js"></script>
+</body>
+</html>
