+++ src/search.js
@@ -1,0 +1,10 @@
+import {getApiData} from "./app";
+
+searchCountry()
+
+function searchCountry() {
+
+
+}
+
+console.log("Hello dummy")
